+++ index.html
@@ -232,5 +232,6 @@
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
 
+  <script type="module" src="js/main.js"></script>
 </body>
 </html>
